backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.0
sqlalchemy==2.0.25
pydantic==2.9.0
python-multipart==0.0.17
websockets==13.1
requests==2.32.3
aiofiles==24.1.0
psutil==6.1.0
python-socketio==5.11.4
patool==3.0.1
psycopg2-binary==2.9.10
python-dotenv==1.0.1
slowapi==0.1.9
joserfc>=1.0.0
passlib>=1.7.4
reportlab>=4.0.0
qrcode[pil]>=7.4
APScheduler>=3.10.0
